Clean up the exec package a little.
